refactor(api): refactor to api modules
refactor api modules

fix api tsconfig

fix tmux script

fix docker configuration
