+++ apps/api/src/schema/modules/Player/index.ts
@@ -1,0 +1,2 @@
+export * from './getPlayerById';
+export * from './Player';
