+++ apps/api/src/schema/modules/Puzzle/Puzzle.ts
@@ -1,6 +1,6 @@
 import { arg, nonNull, objectType } from 'nexus';
 
-import { LanguageArg } from './argTypes';
+import { LanguageArg } from '../types';
 
 export const Puzzle = objectType({
   name: 'Puzzle',
@@ -8,7 +8,6 @@ export const Puzzle = objectType({
     t.nonNull.string('id', { resolve: s => s.id });
     t.nonNull.string('constraint', { resolve: s => s.constraint });
     t.nonNull.string('inputDescription', { resolve: s => s.inputDescription });
-    t.nonNull.string('statement', { resolve: s => s.statement });
     t.nonNull.string('outputDescription', {
       resolve: s => s.outputDescription,
     });
