+++ apps/api/src/schema/modules/Puzzle/PuzzleModeArg.ts
