+++ apps/api/src/schema/modules/Puzzle/index.ts
@@ -1,0 +1,5 @@
+export * from './getPuzzleById';
+export * from './getPuzzleByTitle';
+export * from './getRandomPuzzle';
+export * from './getRandomPuzzleByMode';
+export * from './Puzzle';
