+++ apps/api/src/schema/modules/TestCase/index.ts
@@ -1,0 +1,1 @@
+export * from './TestCase';
