--- apps/api/src/schema/modules/argTypes/index.ts
@@ -1,2 +1,0 @@
-export * from './LanguageArg';
-export * from './PuzzleModeArg';
