+++ apps/api/src/schema/modules/index.ts
@@ -1,4 +1,3 @@
 export * from './Player';
 export * from './Puzzle';
-export * from './Query';
 export * from './TestCase';
