+++ package.json
@@ -6,8 +6,8 @@
   "license": "MIT",
   "scripts": {
     "lint": "nx run-many --target lint --all -- --fix",
-    "start": "nx serve",
-    "start:api": "nx serve api",
+    "start": "nx run playground:serve",
+    "start:api": "nx run api:serve",
     "gen:nexus:api": "nx run api:gen-nexus",
     "build": "nx build",
     "test": "nx test",
@@ -39,15 +39,17 @@
     "@mantine/hooks": "3.6.6",
     "@mantine/next": "3.6.7",
     "@nrwl/next": "13.8.1",
-    "@prisma/client": "3.8.1",
+    "@prisma/client": "3.9.2",
     "apollo-server": "3.4.0",
     "core-js": "^3.0.0",
+    "dotenv-cli": "5.0.0",
     "fp-ts": "2.11.5",
     "fp-ts-contrib": "0.1.26",
     "graphql": "15.6.1",
     "graphql-helix": "1.9.1",
     "io-ts": "2.2.16",
     "io-ts-types": "0.5.16",
+    "newtype-ts": "0.3.5",
     "next": "12.0.7",
     "nexus": "1.1.0",
     "ramda": "0.27.1",
@@ -60,7 +62,7 @@
   "devDependencies": {
     "@babel/core": "7.17.0",
     "@babel/preset-typescript": "7.16.7",
-    "@fullstacksjs/eslint-config": "8.1.0",
+    "@fullstacksjs/eslint-config": "8.1.1",
     "@fullstacksjs/tsconfig": "0.6.0",
     "@nrwl/cli": "13.7.3",
     "@nrwl/cypress": "13.7.3",
@@ -103,7 +105,7 @@
     "jest": "27.5.0",
     "lint-staged": "12.3.3",
     "prettier": "2.5.1",
-    "prisma": "3.8.1",
+    "prisma": "3.9.2",
     "react-test-renderer": "17.0.2",
     "ts-jest": "27.1.3",
     "tsconfig-paths": "3.12.0",
