+++ scripts/dev/tmux
@@ -7,7 +7,7 @@ type tmux >/dev/null 2>&1 || {
 
 SESSION_NAME="codeit"
 COMPOSE="docker-compose -f ./docker-compose.base.yml -f ./docker-compose.dev.yml"
-scripts=("$COMPOSE up api" , "$COMPOSE up postgres" , "npm run studio")
+scripts=("$COMPOSE up api" "$COMPOSE up postgres" "npm run start" "npm run studio")
 
 tmux has-session -t $SESSION_NAME &>/dev/null
 if [ $? != 0 ]; then
@@ -16,12 +16,13 @@ if [ $? != 0 ]; then
       tmux new-session -s $SESSION_NAME -n $SESSION_NAME -d
       INITED=1
     else
-      tmux split-window -t $SESSION_NAME:0 -h -p 33
+      tmux split-window -t $SESSION_NAME:$SESSION_NAME -h -p 33
     fi
-    tmux send-keys -t $SESSION_NAME:0 "$script" C-m
+    echo $script
+    tmux send-keys -t $SESSION_NAME:$SESSION_NAME "$script" C-m
   done
 
   tmux select-layout even-horizontal
 fi
 
-tmux attach-session -t $SESSION_NAME:0
+tmux attach-session -t $SESSION_NAME
